+++ opencc_purepy/union_cache.py
@@ -15,6 +15,8 @@ class UnionKey(Enum):
     TwPhrasesRevOnly = auto()
     TwRevPair = auto()
     Tw2SpR1TwRevTriple = auto()
+    S2HkpR2HkTriple = auto()
+    Hk2SpR1HkRevTriple = auto()
     HkVariantsPair = auto()
     HkRevPair = auto()
     JpVariantsOnly = auto()
@@ -75,6 +77,10 @@ class UnionCache:
             return [get_slot("tw_variants_rev_phrases"), get_slot("tw_variants_rev")]
         if key is UnionKey.Tw2SpR1TwRevTriple:
             return [get_slot("tw_phrases_rev"), get_slot("tw_variants_rev_phrases"), get_slot("tw_variants_rev")]
+        if key is UnionKey.S2HkpR2HkTriple:
+            return [get_slot("hk_phrases"), get_slot("hk_variants_phrases"), get_slot("hk_variants")]
+        if key is UnionKey.Hk2SpR1HkRevTriple:
+            return [get_slot("hk_phrases_rev"), get_slot("hk_variants_rev_phrases"), get_slot("hk_variants_rev")]
         if key is UnionKey.HkVariantsPair:
             return [get_slot("hk_variants_phrases"), get_slot("hk_variants")]
         if key is UnionKey.HkRevPair: